[example] [refactor] Rename some old examples to avoid confusing people (#1149)
* [skip ci] [example] rename some old example names to not confuse people

* [skip ci] rename to mnist_download_data.py, mnist.py become main

* [skip ci] enforce code format

* [skip ci] misc/minimal.py
