+++ examples/gui_image_io.py
