+++ examples/ml_minimization_trivial.py
