--- examples/pytorch_tensor_ad.py
@@ -1,1 +1,0 @@
-# Moved to test/python/test_torch_ad
