--- examples/pytorch_tensor_io.py
@@ -1,1 +1,0 @@
-# Moved to test/python/test_torch_io
