+++ examples/taichi_bitmasked.py
