+++ misc/copyright.py
