+++ misc/minimal.py
